refactor file folders
